Event/Adapter Change
Using an abstraction of BufferedImage (SnippedImage) added additional code with no real benefit.
This has now been changed to BufferedImage.

Added a SaveAdapter that saves the BufferedImage using the current date and time.

The next change should add an event that deals with catching an error message. (Add throws to OnScreenshot in interface and then accept a string to display onError function)